+++ mobile/android/android-components/docs/changelog.md
@@ -15,6 +15,9 @@ permalink: /changelog/
 * **browser-icons**
   * Fixed issue [#7142](https://github.com/mozilla-mobile/android-components/issues/7142)
 
+* **feature-downloads**
+  * On devices older than Q we were not adding downloads to the system download database for more information see [#7230](https://github.com/mozilla-mobile/android-components/issues/7230)
+
 * **browser-storage-sync**
   * Added `getTopFrecentSites` to `PlacesHistoryStorage`, which returns a list of the top frecent site infos
     sorted by most to least frecent.